requirements.txt
Django>=3.2,<4.0
psycopg2>=2.8,<3.0
gunicorn==21.2.0
djangorestframework>=3.12,<3.13
dj-database-url
debugpy==1.5.1
requests
django-redis==5.0.0
redis==5.0.1
